# Allow compressor to repeat Huffman tables